Factoring out loading behavior of MOS
