+++ SpiceSharp/Components/Semiconductors/MOS/Level3/LoadBehavior.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using SpiceSharp.Algebra;
 using SpiceSharp.Attributes;
 using SpiceSharp.Behaviors;
 using SpiceSharp.Simulations;
@@ -9,10 +8,10 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
     /// <summary>
     /// General behavior for a <see cref="Mosfet3"/>
     /// </summary>
-    public class LoadBehavior : BaseLoadBehavior, IConnectedBehavior
+    public class LoadBehavior : Common.LoadBehavior
     {
         /// <summary>
-        /// Necessary behaviors
+        /// Necessary behaviors and parameters
         /// </summary>
         private BaseParameters _bp;
         private ModelBaseParameters _mbp;
@@ -21,71 +20,12 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
         private BaseConfiguration _baseConfig;
 
         /// <summary>
-        /// Shared variables
+        /// Shared parameters
         /// </summary>
         [ParameterName("von"), ParameterInfo("Turn-on voltage")]
         public double Von { get; protected set; }
         [ParameterName("vdsat"), ParameterInfo("Saturation drain voltage")]
         public double SaturationVoltageDs { get; protected set; }
-        [ParameterName("id"), ParameterName("cd"), ParameterInfo("Drain current")]
-        public double DrainCurrent { get; protected set; }
-        [ParameterName("ibs"), ParameterInfo("B-S junction current")]
-        public double BsCurrent { get; protected set; }
-        [ParameterName("ibd"), ParameterInfo("B-D junction current")]
-        public double BdCurrent { get; protected set; }
-        [ParameterName("gmb"), ParameterName("gmbs"), ParameterInfo("Bulk-Source transconductance")]
-        public double TransconductanceBs { get; protected set; }
-        [ParameterName("gm"), ParameterInfo("Transconductance")]
-        public double Transconductance { get; protected set; }
-        [ParameterName("gds"), ParameterInfo("Drain-Source conductance")]
-        public double CondDs { get; protected set; }
-        [ParameterName("gbd"), ParameterInfo("Bulk-Drain conductance")]
-        public double CondBd { get; protected set; }
-        [ParameterName("gbs"), ParameterInfo("Bulk-Source conductance")]
-        public double CondBs { get; protected set; }
-
-        /// <summary>
-        /// Extra variables
-        /// </summary>
-        public int Mode { get; protected set; }
-        public double VoltageBd { get; protected set; }
-        public double VoltageBs { get; protected set; }
-        public double VoltageGs { get; protected set; }
-        public double VoltageDs { get; protected set; }
-
-        /// <summary>
-        /// Nodes
-        /// </summary>
-        private int _drainNode, _gateNode, _sourceNode, _bulkNode;
-        [ParameterName("dnodeprime"), ParameterInfo("Number of protected drain node")]
-        public int DrainNodePrime { get; protected set; }
-        [ParameterName("snodeprime"), ParameterInfo("Number of protected source node")]
-        public int SourceNodePrime { get; protected set; }
-        protected MatrixElement<double> DrainDrainPtr { get; private set; }
-        protected MatrixElement<double> GateGatePtr { get; private set; }
-        protected MatrixElement<double> SourceSourcePtr { get; private set; }
-        protected MatrixElement<double> BulkBulkPtr { get; private set; }
-        protected MatrixElement<double> DrainPrimeDrainPrimePtr { get; private set; }
-        protected MatrixElement<double> SourcePrimeSourcePrimePtr { get; private set; }
-        protected MatrixElement<double> DrainDrainPrimePtr { get; private set; }
-        protected MatrixElement<double> GateBulkPtr { get; private set; }
-        protected MatrixElement<double> GateDrainPrimePtr { get; private set; }
-        protected MatrixElement<double> GateSourcePrimePtr { get; private set; }
-        protected MatrixElement<double> SourceSourcePrimePtr { get; private set; }
-        protected MatrixElement<double> BulkDrainPrimePtr { get; private set; }
-        protected MatrixElement<double> BulkSourcePrimePtr { get; private set; }
-        protected MatrixElement<double> DrainPrimeSourcePrimePtr { get; private set; }
-        protected MatrixElement<double> DrainPrimeDrainPtr { get; private set; }
-        protected MatrixElement<double> BulkGatePtr { get; private set; }
-        protected MatrixElement<double> DrainPrimeGatePtr { get; private set; }
-        protected MatrixElement<double> SourcePrimeGatePtr { get; private set; }
-        protected MatrixElement<double> SourcePrimeSourcePtr { get; private set; }
-        protected MatrixElement<double> DrainPrimeBulkPtr { get; private set; }
-        protected MatrixElement<double> SourcePrimeBulkPtr { get; private set; }
-        protected MatrixElement<double> SourcePrimeDrainPrimePtr { get; private set; }
-        protected VectorElement<double> BulkPtr { get; private set; }
-        protected VectorElement<double> DrainPrimePtr { get; private set; }
-        protected VectorElement<double> SourcePrimePtr { get; private set; }
 
         /// <summary>
         /// Constructor
@@ -102,6 +42,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
         {
             if (provider == null)
                 throw new ArgumentNullException(nameof(provider));
+            base.Setup(simulation, provider);
 
             // Get configurations
             _baseConfig = simulation.Configurations.Get<BaseConfiguration>();
@@ -114,111 +55,25 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
             _temp = provider.GetBehavior<TemperatureBehavior>();
             _modeltemp = provider.GetBehavior<ModelTemperatureBehavior>("model");
 
-            // Initialize some variable
+            // Reset
             SaturationVoltageDs = 0;
             Von = 0;
             Mode = 1;
         }
 
         /// <summary>
-        /// Connect
-        /// </summary>
-        /// <param name="pins">Pins</param>
-        public void Connect(params int[] pins)
-        {
-            if (pins == null)
-                throw new ArgumentNullException(nameof(pins));
-            if (pins.Length != 4)
-                throw new CircuitException("Pin count mismatch: 4 pins expected, {0} given".FormatString(pins.Length));
-            _drainNode = pins[0];
-            _gateNode = pins[1];
-            _sourceNode = pins[2];
-            _bulkNode = pins[3];
-        }
-
-        /// <summary>
-        /// Gets matrix pointers
-        /// </summary>
-        /// <param name="variables">Variables</param>
-        /// <param name="solver">Matrix</param>
-        public override void GetEquationPointers(VariableSet variables, Solver<double> solver)
-        {
-            if (variables == null)
-                throw new ArgumentNullException(nameof(variables));
-            if (solver == null)
-                throw new ArgumentNullException(nameof(solver));
-
-            // Add a series drain node if necessary
-            if (_mbp.DrainResistance > 0 || _mbp.SheetResistance > 0 && _bp.DrainSquares > 0)
-                DrainNodePrime = variables.Create(Name.Combine("drain")).Index;
-            else
-                DrainNodePrime = _drainNode;
-
-            // Add a series source node if necessary
-            if (_mbp.SourceResistance > 0 || _mbp.SheetResistance > 0 && _bp.SourceSquares > 0)
-                SourceNodePrime = variables.Create(Name.Combine("source")).Index;
-            else
-                SourceNodePrime = _sourceNode;
-
-            // Get matrix elements
-            DrainDrainPtr = solver.GetMatrixElement(_drainNode, _drainNode);
-            GateGatePtr = solver.GetMatrixElement(_gateNode, _gateNode);
-            SourceSourcePtr = solver.GetMatrixElement(_sourceNode, _sourceNode);
-            BulkBulkPtr = solver.GetMatrixElement(_bulkNode, _bulkNode);
-            DrainPrimeDrainPrimePtr = solver.GetMatrixElement(DrainNodePrime, DrainNodePrime);
-            SourcePrimeSourcePrimePtr = solver.GetMatrixElement(SourceNodePrime, SourceNodePrime);
-            DrainDrainPrimePtr = solver.GetMatrixElement(_drainNode, DrainNodePrime);
-            GateBulkPtr = solver.GetMatrixElement(_gateNode, _bulkNode);
-            GateDrainPrimePtr = solver.GetMatrixElement(_gateNode, DrainNodePrime);
-            GateSourcePrimePtr = solver.GetMatrixElement(_gateNode, SourceNodePrime);
-            SourceSourcePrimePtr = solver.GetMatrixElement(_sourceNode, SourceNodePrime);
-            BulkDrainPrimePtr = solver.GetMatrixElement(_bulkNode, DrainNodePrime);
-            BulkSourcePrimePtr = solver.GetMatrixElement(_bulkNode, SourceNodePrime);
-            DrainPrimeSourcePrimePtr = solver.GetMatrixElement(DrainNodePrime, SourceNodePrime);
-            DrainPrimeDrainPtr = solver.GetMatrixElement(DrainNodePrime, _drainNode);
-            BulkGatePtr = solver.GetMatrixElement(_bulkNode, _gateNode);
-            DrainPrimeGatePtr = solver.GetMatrixElement(DrainNodePrime, _gateNode);
-            SourcePrimeGatePtr = solver.GetMatrixElement(SourceNodePrime, _gateNode);
-            SourcePrimeSourcePtr = solver.GetMatrixElement(SourceNodePrime, _sourceNode);
-            DrainPrimeBulkPtr = solver.GetMatrixElement(DrainNodePrime, _bulkNode);
-            SourcePrimeBulkPtr = solver.GetMatrixElement(SourceNodePrime, _bulkNode);
-            SourcePrimeDrainPrimePtr = solver.GetMatrixElement(SourceNodePrime, DrainNodePrime);
-
-            // Get rhs elements
-            BulkPtr = solver.GetRhsElement(_bulkNode);
-            DrainPrimePtr = solver.GetRhsElement(DrainNodePrime);
-            SourcePrimePtr = solver.GetRhsElement(SourceNodePrime);
-        }
-
-        /// <summary>
-        /// Unsetup
+        /// Unsetup the behavior
         /// </summary>
         /// <param name="simulation"></param>
         public override void Unsetup(Simulation simulation)
         {
-            // Remove references
-            DrainDrainPtr = null;
-            GateGatePtr = null;
-            SourceSourcePtr = null;
-            BulkBulkPtr = null;
-            DrainPrimeDrainPrimePtr = null;
-            SourcePrimeSourcePrimePtr = null;
-            DrainDrainPrimePtr = null;
-            GateBulkPtr = null;
-            GateDrainPrimePtr = null;
-            GateSourcePrimePtr = null;
-            SourceSourcePrimePtr = null;
-            BulkDrainPrimePtr = null;
-            BulkSourcePrimePtr = null;
-            DrainPrimeSourcePrimePtr = null;
-            DrainPrimeDrainPtr = null;
-            BulkGatePtr = null;
-            DrainPrimeGatePtr = null;
-            SourcePrimeGatePtr = null;
-            SourcePrimeSourcePtr = null;
-            DrainPrimeBulkPtr = null;
-            SourcePrimeBulkPtr = null;
-            SourcePrimeDrainPrimePtr = null;
+            _baseConfig = null;
+            _bp = null;
+            _mbp = null;
+            _temp = null;
+            _modeltemp = null;
+
+            base.Unsetup(simulation);
         }
 
         /// <summary>
@@ -234,23 +89,14 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
             double drainSatCur, sourceSatCur,
                 vgs, vds, vbs, vbd, vgd;
             double von;
-            double vdsat,
-                cdrain,
-                cdreq;
+            double vdsat, cdrain = 0.0, cdreq;
             int xnrm, xrev;
 
             var vt = Circuit.KOverQ * _bp.Temperature;
             var check = 1;
 
-            /* DETAILPROF */
-
-            /* first, we compute a few useful values - these could be
-			* pre - computed, but for historical reasons are still done
-			* here.  They may be moved at the expense of instance size
-			*/
-
             var effectiveLength = _bp.Length - 2 * _mbp.LateralDiffusion;
-            if (_temp.TempSaturationCurrentDensity.Equals(0.0) || _bp.DrainArea.Value.Equals(0.0) || _bp.SourceArea.Value.Equals(0.0))
+            if (_temp.TempSaturationCurrentDensity.Equals(0) || _bp.DrainArea.Value <= 0 || _bp.SourceArea.Value <= 0)
             {
                 drainSatCur = _temp.TempSaturationCurrent;
                 sourceSatCur = _temp.TempSaturationCurrent;
@@ -264,41 +110,26 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
             var beta = _temp.TempTransconductance * _bp.Width / effectiveLength;
             var oxideCap = _mbp.OxideCapFactor * effectiveLength * _bp.Width;
 
-            /* DETAILPROF */
-
-            /* 
-			* ok - now to do the start - up operations
-			* 
-			* we must get values for vbs, vds, and vgs from somewhere
-			* so we either predict them or recover them from last iteration
-			* These are the two most common cases - either a prediction
-			* step or the general iteration step and they
-			* share some code, so we put them first - others later on
-			*/
-
             if (state.Init == InitializationModes.Float || (simulation is TimeSimulation tsim && tsim.Method.BaseTime.Equals(0.0)) ||
                 state.Init == InitializationModes.Fix && !_bp.Off)
             {
                 // General iteration
-                vbs = _mbp.MosfetType * (state.Solution[_bulkNode] - state.Solution[SourceNodePrime]);
-                vgs = _mbp.MosfetType * (state.Solution[_gateNode] - state.Solution[SourceNodePrime]);
+                vbs = _mbp.MosfetType * (state.Solution[BulkNode] - state.Solution[SourceNodePrime]);
+                vgs = _mbp.MosfetType * (state.Solution[GateNode] - state.Solution[SourceNodePrime]);
                 vds = _mbp.MosfetType * (state.Solution[DrainNodePrime] - state.Solution[SourceNodePrime]);
 
-                /* now some common crunching for some more useful quantities */
-                /* DETAILPROF */
-
+                // now some common crunching for some more useful quantities
                 vbd = vbs - vds;
                 vgd = vgs - vds;
                 var vgdo = VoltageGs - VoltageDs;
                 von = _mbp.MosfetType * Von;
 
-                /* 
-				* limiting
-				* we want to keep device voltages from changing
-				* so fast that the exponentials churn out overflows
-				* and similar rudeness
-				*/
-
+                /*
+				 * limiting
+				 * we want to keep device voltages from changing
+				 * so fast that the exponentials churn out overflows
+				 * and similar rudeness
+				 */
                 // NOTE: Spice 3f5 does not write out Vgs during DC analysis, so DEVfetlim may give different results in Spice 3f5
                 if (VoltageDs >= 0)
                 {
@@ -322,16 +153,13 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     vbd = Transistor.LimitJunction(vbd, VoltageBd, vt, _temp.DrainVCritical, out check);
                     vbs = vbd + vds;
                 }
-                /* NODELIMITING */
             }
             else
             {
-                /* DETAILPROF */
                 /* ok - not one of the simple cases, so we have to
-				* look at all of the possibilities for why we were
-				* called.  We still just initialize the three voltages
-				*/
-
+				 * look at all of the possibilities for why we were
+				 * called.  We still just initialize the three voltages
+				 */
                 if (state.Init == InitializationModes.Junction && !_bp.Off)
                 {
                     vds = _mbp.MosfetType * _bp.InitialVoltageDs;
@@ -339,7 +167,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     vbs = _mbp.MosfetType * _bp.InitialVoltageBs;
 
                     // TODO: At some point, check what this is supposed to do
-                    if (vds.Equals(0.0) && vgs.Equals(0.0) && vbs.Equals(0.0) && (state.UseDc || !state.UseIc))
+                    if (vds.Equals(0) && vgs.Equals(0) && vbs.Equals(0) && (state.UseDc || !state.UseIc))
                     {
                         vbs = -1;
                         vgs = _mbp.MosfetType * _temp.TempVt0;
@@ -352,19 +180,18 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 }
             }
 
-            /* DETAILPROF */
-            /* 
-			* now all the preliminaries are over - we can start doing the
-			* real work
-			*/
+            /*
+			 * now all the preliminaries are over - we can start doing the
+			 * real work
+			 */
             vbd = vbs - vds;
             vgd = vgs - vds;
 
-            /* 
-			* bulk - source and bulk - drain diodes
-			* here we just evaluate the ideal diode current and the
-			* corresponding derivative (conductance).
-			*/
+            /*
+			 * bulk - source and bulk - drain diodes
+			 * here we just evaluate the ideal diode current and the
+			 * corresponding derivative (conductance).
+			 */
             if (vbs <= 0)
             {
                 CondBs = sourceSatCur / vt;
@@ -391,41 +218,41 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
             }
 
             /* now to determine whether the user was able to correctly
-			* identify the source and drain of his device
-			*/
+			 * identify the source and drain of his device
+			 */
             if (vds >= 0)
             {
-                /* normal mode */
+                // normal mode
                 Mode = 1;
             }
             else
             {
-                /* inverse mode */
+                // inverse mode
                 Mode = -1;
             }
 
             /* DETAILPROF */
             {
-                /* 
-				* subroutine moseq3(vds, vbs, vgs, gm, gds, gmbs, 
+                /*
+				* subroutine moseq3(vds, vbs, vgs, gm, gds, gmbs,
 				* qg, qc, qb, cggb, cgdb, cgsb, cbgb, cbdb, cbsb)
 				*/
 
-                /* 
+                /*
 				* this routine evaluates the drain current, its derivatives and
 				* the charges associated with the gate, channel and bulk
 				* for mosfets based on semi - empirical equations
 				*/
 
-                /* 
-				common / mosarg / vto, beta, gamma, phi, phib, cox, xnsub, xnfs, xd, xj, xld, 
-				1   xlamda, uo, uexp, vbp, utra, vmax, xneff, xl, xw, vbi, von, vdsat, qspof, 
+                /*
+				common / mosarg / vto, beta, gamma, phi, phib, cox, xnsub, xnfs, xd, xj, xld,
+				1   xlamda, uo, uexp, vbp, utra, vmax, xneff, xl, xw, vbi, von, vdsat, qspof,
 				2   beta0, beta1, cdrain, xqco, xqc, fnarrw, fshort, lev
-				common / status / omega, time, delta, delold(7), ag(7), vt, xni, egfet, 
-				1   xmu, sfactr, mode, modedc, icalc, initf, method, iord, maxord, noncon, 
+				common / status / omega, time, delta, delold(7), ag(7), vt, xni, egfet,
+				1   xmu, sfactr, mode, modedc, icalc, initf, method, iord, maxord, noncon,
 				2   iterno, itemno, nosolv, modac, ipiv, ivmflg, ipostp, iscrch, iofile
-				common / knstnt / twopi, xlog2, xlog10, root2, rad, boltz, charge, ctok, 
-				1   gmin, reltol, abstol, vntol, trtol, chgtol, eps0, epssil, epsox, 
+				common / knstnt / twopi, xlog2, xlog10, root2, rad, boltz, charge, ctok,
+				1   gmin, reltol, abstol, vntol, trtol, chgtol, eps0, epssil, epsox,
 				2   pivtol, pivrel
 				*/
 
@@ -450,18 +277,18 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     gds0 = 0.0;
                 double fshort;
 
-                /* 
+                /*
 				* bypasses the computation of charges
 				*/
 
-                /* 
+                /*
 				* reference cdrain equations to source and
 				* charge equations to bulk
 				*/
                 vdsat = 0.0;
                 var oneoverxl = 1.0 / effectiveLength;
                 var eta = _mbp.Eta * 8.15e-22 / (_mbp.OxideCapFactor * effectiveLength * effectiveLength * effectiveLength);
-                /* 
+                /*
 				* .....square root term
 				*/
                 if ((Mode == 1 ? vbs : vbd) <= 0.0)
@@ -478,7 +305,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     phibs = sqphbs * sqphbs;
                     dsqdvb = -phibs / (sqphs3 + sqphs3);
                 }
-                /* 
+                /*
 				 * .....short channel effect factor
 				 */
                 if (_mbp.JunctionDepth > 0 && _modeltemp.CoefficientDepletionLayerWidth > 0.0)
@@ -503,7 +330,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     fshort = 1.0;
                     dfsdvb = 0.0;
                 }
-                /* 
+                /*
 				 * .....body effect
 				 */
                 var gammas = _mbp.Gamma * fshort;
@@ -513,17 +340,17 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 var dfbdvb = -fbodys * dsqdvb / sqphbs + fbodys * dfsdvb / fshort;
                 var qbonco = gammas * sqphbs + _mbp.NarrowFactor * phibs / _bp.Width;
                 var dqbdvb = gammas * dsqdvb + _mbp.Gamma * dfsdvb * sqphbs - _mbp.NarrowFactor / _bp.Width;
-                /* 
+                /*
 				 * .....static feedback effect
 				 */
                 var vbix = _temp.TempVoltageBi * _mbp.MosfetType - eta * (Mode * vds);
-                /* 
+                /*
 				 * .....threshold voltage
 				 */
                 var vth = vbix + qbonco;
                 var dvtdvd = -eta;
                 var dvtdvb = dqbdvb;
-                /* 
+                /*
 				 * .....joint weak inversion and strong inversion
 				 */
                 von = vth;
@@ -540,7 +367,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 }
                 else
                 {
-                    /* 
+                    /*
 					 * .....cutoff region
 					 */
                     if ((Mode == 1 ? vgs : vgd) <= von)
@@ -552,11 +379,11 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                         goto innerline1000;
                     }
                 }
-                /* 
+                /*
 				 * .....device is on
 				 */
                 var vgsx = Math.Max(Mode == 1 ? vgs : vgd, von);
-                /* 
+                /*
 				 * .....mobility modulation by gate voltage
 				 */
                 var onfg = 1.0 + _mbp.Theta * (vgsx - vth);
@@ -565,7 +392,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 var dfgdvg = -_mbp.Theta * fgate * fgate;
                 var dfgdvd = -dfgdvg * dvtdvd;
                 var dfgdvb = -dfgdvg * dvtdvb;
-                /* 
+                /*
 				 * .....saturation voltage
 				 */
                 vdsat = (vgsx - vth) * onfbdy;
@@ -587,7 +414,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     dvsdvd = -dvsdvg * dvtdvd;
                     dvsdvb = -dvsdvg * dvtdvb - arga * dvsdga * dfbdvb;
                 }
-                /* 
+                /*
 				 * .....current factors in linear region
 				 */
                 var vdsx = Math.Min(Mode * vds, vdsat);
@@ -595,14 +422,14 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     goto line900;
                 var cdo = vgsx - vth - 0.5 * (1.0 + fbody) * vdsx;
                 var dcodvb = -dvtdvb - 0.5 * dfbdvb * vdsx;
-                /* 
+                /*
 				 * .....normalized drain current
 				 */
                 var cdnorm = cdo * vdsx;
                 Transconductance = vdsx;
                 CondDs = vgsx - vth - (1.0 + fbody + dvtdvd) * vdsx;
                 TransconductanceBs = dcodvb * vdsx;
-                /* 
+                /*
 				 * .....drain current without velocity saturation effect
 				 */
                 var cd1 = beta * cdnorm;
@@ -611,7 +438,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 Transconductance = beta * Transconductance + dfgdvg * cd1;
                 CondDs = beta * CondDs + dfgdvd * cd1;
                 TransconductanceBs = beta * TransconductanceBs;
-                /* 
+                /*
 				 * .....velocity saturation factor
 				 */
                 if (_mbp.MaxDriftVelocity > 0.0)
@@ -622,7 +449,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     dfddvg = -dfgdvg * arga;
                     dfddvd = -dfgdvd * arga - fd2 * onvdsc;
                     dfddvb = -dfgdvb * arga;
-                    /* 
+                    /*
 					 * .....drain current
 					 */
                     Transconductance = fdrain * Transconductance + dfddvg * cdrain;
@@ -630,7 +457,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     TransconductanceBs = fdrain * TransconductanceBs + dfddvb * cdrain;
                     cdrain = fdrain * cdrain;
                 }
-                /* 
+                /*
 				 * .....channel length modulation
 				 */
                 if (Mode * vds <= vdsat) goto line700;
@@ -670,7 +497,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 ddldvg = 0.0;
                 ddldvd = -dldvd;
                 ddldvb = 0.0;
-                /* 
+                /*
 				 * .....punch through approximation
 				 */
                 line520:
@@ -683,7 +510,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     ddldvb = ddldvb * arga;
                     dldvd = dldvd * arga;
                 }
-                /* 
+                /*
 				 * .....saturation region
 				 */
                 var dlonxl = delxl * oneoverxl;
@@ -696,13 +523,13 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                 Transconductance = Transconductance + gds0 * dvsdvg;
                 TransconductanceBs = TransconductanceBs + gds0 * dvsdvb;
                 CondDs = gds0 * dvsdvd + diddl * dldvd;
-                /* 
+                /*
 				 * .....finish strong inversion case
 				 */
                 line700:
                 if ((Mode == 1 ? vgs : vgd) < von)
                 {
-                    /* 
+                    /*
 					 * .....weak inversion
 					 */
                     var onxn = 1.0 / xn;
@@ -719,11 +546,11 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     CondDs = CondDs * wfact + (gms - gmw) * dvodvd;
                     TransconductanceBs = TransconductanceBs * wfact + (gms - gmw) * dvodvb - gmw * ((Mode == 1 ? vgs : vgd) - von) * onxn * dxndvb;
                 }
-                /* 
+                /*
 				 * .....charge computation
 				 */
                 goto innerline1000;
-                /* 
+                /*
 				 * .....special case of vds = 0.0d0 */
                 line900:
                 beta = beta * fgate;
@@ -736,24 +563,19 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     CondDs *= Math.Exp(((Mode == 1 ? vgs : vgd) - von) / (vt * xn));
                 }
                 innerline1000:;
-                /* 
+                /*
 				 * .....done
 				 */
             }
 
-            /* DETAILPROF */
-
-            /* now deal with n vs p polarity */
-
             Von = _mbp.MosfetType * von;
             SaturationVoltageDs = _mbp.MosfetType * vdsat;
-            /* line 490 */
-            /* 
-			* COMPUTE EQUIVALENT DRAIN CURRENT SOURCE
-			*/
+            /*
+			 * COMPUTE EQUIVALENT DRAIN CURRENT SOURCE
+			 */
             DrainCurrent = Mode * cdrain - BdCurrent;
 
-            /* 
+            /*
 			 * check convergence
 			 */
             if (!_bp.Off || state.Init != InitializationModes.Fix)
@@ -762,21 +584,12 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
                     state.IsConvergent = false;
             }
 
-            /* DETAILPROF */
-
-            /* save things away for next time */
             VoltageBs = vbs;
             VoltageBd = vbd;
             VoltageGs = vgs;
             VoltageDs = vds;
-            /* DETAILPROF */
 
-            /* 
-			 * meyer's capacitor model
-			 */
-
-            /* DETAILPROF */
-            /* 
+            /*
 			 * load current vector
 			 */
             var ceqbs = _mbp.MosfetType * (BsCurrent - (CondBs - _baseConfig.Gmin) * vbs);
@@ -815,65 +628,6 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
             SourcePrimeSourcePtr.Value += -_temp.SourceConductance;
             SourcePrimeBulkPtr.Value += -CondBs - (xnrm - xrev) * TransconductanceBs;
             SourcePrimeDrainPrimePtr.Value += -CondDs - xrev * (Transconductance + TransconductanceBs);
-        }
-
-        /// <summary>
-        /// Check convergence
-        /// </summary>
-        /// <param name="simulation">Base simulation</param>
-        /// <returns></returns>
-        public override bool IsConvergent(BaseSimulation simulation)
-        {
-			if (simulation == null)
-				throw new ArgumentNullException(nameof(simulation));
-
-            var state = simulation.RealState;
-            double cdhat;
-
-            var vbs = _mbp.MosfetType * (state.Solution[_bulkNode] - state.Solution[SourceNodePrime]);
-            var vgs = _mbp.MosfetType * (state.Solution[_gateNode] - state.Solution[SourceNodePrime]);
-            var vds = _mbp.MosfetType * (state.Solution[DrainNodePrime] - state.Solution[SourceNodePrime]);
-            var vbd = vbs - vds;
-            var vgd = vgs - vds;
-            var vgdo = VoltageGs - VoltageDs;
-            var delvbs = vbs - VoltageBs;
-            var delvbd = vbd - VoltageBd;
-            var delvgs = vgs - VoltageGs;
-            var delvds = vds - VoltageDs;
-            var delvgd = vgd - vgdo;
-
-            /* these are needed for convergence testing */
-
-            if (Mode >= 0)
-            {
-                cdhat = DrainCurrent - CondBd * delvbd + TransconductanceBs * delvbs +
-                    Transconductance * delvgs + CondDs * delvds;
-            }
-            else
-            {
-                cdhat = DrainCurrent - (CondBd - TransconductanceBs) * delvbd -
-                    Transconductance * delvgd + CondDs * delvds;
-            }
-            var cbhat = BsCurrent + BdCurrent + CondBd * delvbd + CondBs * delvbs;
-
-            /*
-             *  check convergence
-             */
-            var tol = _baseConfig.RelativeTolerance * Math.Max(Math.Abs(cdhat), Math.Abs(DrainCurrent)) + _baseConfig.AbsoluteTolerance;
-            if (Math.Abs(cdhat - DrainCurrent) >= tol)
-            {
-                state.IsConvergent = false;
-                return false;
-            }
-
-            tol = _baseConfig.RelativeTolerance * Math.Max(Math.Abs(cbhat), Math.Abs(BsCurrent + BdCurrent)) + _baseConfig.AbsoluteTolerance;
-            if (Math.Abs(cbhat - (BsCurrent + BdCurrent)) > tol)
-            {
-                state.IsConvergent = false;
-                return false;
-            }
-
-            return true;
         }
     }
 }
